+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [7.0.1] - 2020-05-18
+
+### Fixed
+
 - #1473, Fix overloaded computed columns on RPC - @wolfgangwalther
 - #1471, Fix POST, PATCH, DELETE with ?select= and return=minimal and PATCH with empty body - @wolfgangwalther
 - #1500, Fix missing `openapi-server-proxy-uri` config option - @steve-chavez